p3_airflow/requirements.txt
apache-airflow-providers-google==6.2.0
apache-airflow[google]==2.2.3
pyarrow
